# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SS Milhas API",
    "version": "1.0.0",
    "description": "API completa do Sistema de Gestão de Milhas Aéreas",
    "contact": {
      "name": "Suporte SS Milhas",
      "email": "suporte@ssmilhas.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://ss-milhas.vercel.app/api",
      "description": "Servidor de Produção"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Fazer login",
        "description": "Autentica um usuário e retorna um token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "senha"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@example.com"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "example": "SenhaSegura123!@#"
                  },
                  "lembrarMe": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimit"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registrar novo usuário",
        "description": "Cria uma nova conta de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "email", "senha"],
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "João Silva"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "joao@example.com"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "example": "SenhaSegura123!@#"
                  },
                  "telefone": {
                    "type": "string",
                    "example": "11988888888"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "12345678901"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimit"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "manager", "user"]
          },
          "status": {
            "type": "string",
            "enum": ["ativo", "inativo", "suspenso"]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "permissions": {
            "type": "object",
            "properties": {
              "financeiro": {
                "type": "boolean"
              },
              "valores": {
                "type": "boolean"
              },
              "relatorios": {
                "type": "boolean"
              },
              "monitoramento": {
                "type": "boolean"
              },
              "cadastros": {
                "type": "boolean"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": false
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Não autorizado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimit": {
        "description": "Limite de requisições excedido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}